backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.8.0
openai==1.3.8
python-multipart==0.0.6
unstructured==0.12.5
pillow==10.1.0
pypdf==3.17.4
python-docx==0.8.11
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0
pytz==2023.3.post1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
tiktoken==0.5.2
sentence-transformers==2.2.2